requirements.txt
Flask==3.0.2
SQLAlchemy>=2.0.38
requests==2.31.0
feedparser==6.0.11
beautifulsoup4==4.12.3
python-dateutil==2.8.2
spacy>=3.7.4,<3.8
alembic==1.13.1
psycopg[binary]==3.1.18
prometheus-client==0.20.0
sentry-sdk[flask]==1.45.0
redis==5.0.1
